requirements.txt
Flask==0.11.1
Flask-Cors==2.1.2
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
ruamel-yaml===-VERSION
